Just use ints when passing coordinates around. Using short just added about 1k of extra code!

+++ libs/graphics/graphics.c
@@ -42,22 +42,22 @@
 
 static void graphicsSetPixelDevice(JsGraphics *gfx, int x, int y, unsigned int col);
 
-void graphicsFallbackSetPixel(JsGraphics *gfx, short x, short y, unsigned int col) {
+void graphicsFallbackSetPixel(JsGraphics *gfx, int x, int y, unsigned int col) {
   NOT_USED(gfx);
   NOT_USED(x);
   NOT_USED(y);
   NOT_USED(col);
 }
 
-unsigned int graphicsFallbackGetPixel(JsGraphics *gfx, short x, short y) {
+unsigned int graphicsFallbackGetPixel(JsGraphics *gfx, int x, int y) {
   NOT_USED(gfx);
   NOT_USED(x);
   NOT_USED(y);
   return 0;
 }
 
-void graphicsFallbackFillRect(JsGraphics *gfx, short x1, short y1, short x2, short y2) {
-  short x,y;
+void graphicsFallbackFillRect(JsGraphics *gfx, int x1, int y1, int x2, int y2) {
+  int x,y;
   for (y=y1;y<=y2;y++)
     for (x=x1;x<=x2;x++)
       graphicsSetPixelDevice(gfx,x,y, gfx->data.fgColor);
@@ -68,12 +68,12 @@ void graphicsFallbackScrollX(JsGraphics *gfx, int xdir, int yfrom, int yto) {
   if (xdir<=0) {
     int w = gfx->data.width+xdir;
     for (x=0;x<w;x++)
-      gfx->setPixel(gfx, (short)x,(short)yto,
-          gfx->getPixel(gfx, (short)(x-xdir),(short)yfrom));
+      gfx->setPixel(gfx, (int)x,(int)yto,
+          gfx->getPixel(gfx, (int)(x-xdir),(int)yfrom));
   } else { // >0
     for (x=gfx->data.width-xdir-1;x>=0;x--)
-      gfx->setPixel(gfx, (short)(x+xdir),(short)yto,
-          gfx->getPixel(gfx, (short)x,(short)yfrom));
+      gfx->setPixel(gfx, (int)(x+xdir),(int)yto,
+          gfx->getPixel(gfx, (int)x,(int)yfrom));
   }
 }
 
@@ -189,14 +189,14 @@ size_t graphicsGetMemoryRequired(const JsGraphics *gfx) {
 // ----------------------------------------------------------------------------------------------
 
 // If graphics is flipped or rotated then the coordinates need modifying
-void graphicsToDeviceCoordinates(const JsGraphics *gfx, short *x, short *y) {
+void graphicsToDeviceCoordinates(const JsGraphics *gfx, int *x, int *y) {
   if (gfx->data.flags & JSGRAPHICSFLAGS_SWAP_XY) {
-    short t = *x;
+    int t = *x;
     *x = *y;
     *y = t;
   }
-  if (gfx->data.flags & JSGRAPHICSFLAGS_INVERT_X) *x = (short)(gfx->data.width - (*x+1));
-  if (gfx->data.flags & JSGRAPHICSFLAGS_INVERT_Y) *y = (short)(gfx->data.height - (*y+1));
+  if (gfx->data.flags & JSGRAPHICSFLAGS_INVERT_X) *x = (int)(gfx->data.width - (*x+1));
+  if (gfx->data.flags & JSGRAPHICSFLAGS_INVERT_Y) *y = (int)(gfx->data.height - (*y+1));
 }
 
 // ----------------------------------------------------------------------------------------------
@@ -209,10 +209,10 @@ static void graphicsSetPixelDevice(JsGraphics *gfx, int x, int y, unsigned int c
   if (y < gfx->data.modMinY) gfx->data.modMinY=(short)y;
   if (y > gfx->data.modMaxY) gfx->data.modMaxY=(short)y;
 #endif
-  gfx->setPixel(gfx,(short)x,(short)y,col & (unsigned int)((1L<<gfx->data.bpp)-1));
+  gfx->setPixel(gfx,(int)x,(int)y,col & (unsigned int)((1L<<gfx->data.bpp)-1));
 }
 
-static unsigned int graphicsGetPixelDevice(JsGraphics *gfx, short x, short y) {
+static unsigned int graphicsGetPixelDevice(JsGraphics *gfx, int x, int y) {
   if (x<0 || y<0 || x>=gfx->data.width || y>=gfx->data.height) return 0;
   return gfx->getPixel(gfx, x, y);
 }
@@ -240,26 +240,26 @@ static void graphicsFillRectDevice(JsGraphics *gfx, int x1, int y1, int x2, int 
   if (y2 > gfx->data.modMaxY) gfx->data.modMaxY=(short)y2;
 #endif
   if (x1==x2 && y1==y2) {
-    gfx->setPixel(gfx,(short)x1,(short)y1,gfx->data.fgColor);
+    gfx->setPixel(gfx,(int)x1,(int)y1,gfx->data.fgColor);
     return;
   }
 
-  return gfx->fillRect(gfx, (short)x1, (short)y1, (short)x2, (short)y2);
+  return gfx->fillRect(gfx, (int)x1, (int)y1, (int)x2, (int)y2);
 }
 
 // ----------------------------------------------------------------------------------------------
 
-void graphicsSetPixel(JsGraphics *gfx, short x, short y, unsigned int col) {
+void graphicsSetPixel(JsGraphics *gfx, int x, int y, unsigned int col) {
   graphicsToDeviceCoordinates(gfx, &x, &y);
   graphicsSetPixelDevice(gfx, x, y, col);
 }
 
-unsigned int graphicsGetPixel(JsGraphics *gfx, short x, short y) {
+unsigned int graphicsGetPixel(JsGraphics *gfx, int x, int y) {
   graphicsToDeviceCoordinates(gfx, &x, &y);
   return graphicsGetPixelDevice(gfx, x, y);
 }
 
-void graphicsFillRect(JsGraphics *gfx, short x1, short y1, short x2, short y2) {
+void graphicsFillRect(JsGraphics *gfx, int x1, int y1, int x2, int y2) {
   graphicsToDeviceCoordinates(gfx, &x1, &y1);
   graphicsToDeviceCoordinates(gfx, &x2, &y2);
   graphicsFillRectDevice(gfx, x1, y1, x2, y2);
@@ -268,14 +268,14 @@ void graphicsFillRect(JsGraphics *gfx, short x1, short y1, short x2, short y2) {
 void graphicsClear(JsGraphics *gfx) {
   unsigned int c = gfx->data.fgColor;
   gfx->data.fgColor = gfx->data.bgColor;
-  graphicsFillRectDevice(gfx,0,0,(short)(gfx->data.width-1),(short)(gfx->data.height-1));
+  graphicsFillRectDevice(gfx,0,0,(int)(gfx->data.width-1),(int)(gfx->data.height-1));
   gfx->data.fgColor = c;
 }
 
 // ----------------------------------------------------------------------------------------------
 
 
-void graphicsDrawRect(JsGraphics *gfx, short x1, short y1, short x2, short y2) {
+void graphicsDrawRect(JsGraphics *gfx, int x1, int y1, int x2, int y2) {
   graphicsToDeviceCoordinates(gfx, &x1, &y1);
   graphicsToDeviceCoordinates(gfx, &x2, &y2);
   // rather than writing pixels, we use fillrect - as it is faster
@@ -285,7 +285,7 @@ void graphicsDrawRect(JsGraphics *gfx, short x1, short y1, short x2, short y2) {
   graphicsFillRectDevice(gfx,x1,y2,x1,y1);
 }
 
-void graphicsDrawEllipse(JsGraphics *gfx, short posX1, short posY1, short posX2, short posY2){
+void graphicsDrawEllipse(JsGraphics *gfx, int posX1, int posY1, int posX2, int posY2){
   graphicsToDeviceCoordinates(gfx, &posX1, &posY1);
   graphicsToDeviceCoordinates(gfx, &posX2, &posY2);
   int posX = (posX1+posX2)/2;
@@ -323,7 +323,7 @@ void graphicsDrawEllipse(JsGraphics *gfx, short posX1, short posY1, short posX2,
   }
 }
 
-void graphicsFillEllipse(JsGraphics *gfx, short posX1, short posY1, short posX2, short posY2){
+void graphicsFillEllipse(JsGraphics *gfx, int posX1, int posY1, int posX2, int posY2){
   graphicsToDeviceCoordinates(gfx, &posX1, &posY1);
   graphicsToDeviceCoordinates(gfx, &posX2, &posY2);
   int posX = (posX1+posX2)/2;
@@ -350,15 +350,15 @@ void graphicsFillEllipse(JsGraphics *gfx, short posX1, short posY1, short posX2,
   }
 }
 
-static void graphicsDrawString(JsGraphics *gfx, short x1, short y1, const char *str) {
+static void graphicsDrawString(JsGraphics *gfx, int x1, int y1, const char *str) {
   // no need to modify coordinates as setPixel does that
   while (*str) {
     graphicsDrawChar4x6(gfx,x1,y1,*(str++),1);
-    x1 = (short)(x1 + 4);
+    x1 = (int)(x1 + 4);
   }
 }
 
-void graphicsDrawLine(JsGraphics *gfx, short x1, short y1, short x2, short y2) {
+void graphicsDrawLine(JsGraphics *gfx, int x1, int y1, int x2, int y2) {
   graphicsToDeviceCoordinates(gfx, &x1, &y1);
   graphicsToDeviceCoordinates(gfx, &x2, &y2);
 
@@ -368,26 +368,26 @@ void graphicsDrawLine(JsGraphics *gfx, short x1, short y1, short x2, short y2) {
   if (yl<0) yl=-yl; else if (yl==0) yl=1;
   if (xl > yl) { // longer in X - scan in X
     if (x1>x2) {
-      short t;
+      int t;
       t = x1; x1 = x2; x2 = t;
       t = y1; y1 = y2; y2 = t;
     }
     int pos = (y1<<8) + 128; // rounding!
     int step = ((y2-y1)<<8) / xl;
-    short x;
+    int x;
     for (x=x1;x<=x2;x++) {
       graphicsSetPixelDevice(gfx, x, pos>>8, gfx->data.fgColor);
       pos += step;
     }
   } else {
     if (y1>y2) {
-      short t;
+      int t;
       t = x1; x1 = x2; x2 = t;
       t = y1; y1 = y2; y2 = t;
     }
     int pos = (x1<<8) + 128; // rounding!
     int step = ((x2-x1)<<8) / yl;
-    short y;
+    int y;
     for (y=y1;y<=y2;y++) {
       graphicsSetPixelDevice(gfx, pos>>8, y, gfx->data.fgColor);
       pos += step;
@@ -425,8 +425,8 @@ static inline void graphicsFillPolyCreateScanLines(JsGraphics *gfx, short *minx,
 
 void graphicsFillPoly(JsGraphics *gfx, int points, short *vertices) {
   int i;
-  short miny = (short)(gfx->data.height-1);
-  short maxy = 0;
+  int miny = (int)(gfx->data.height-1);
+  int maxy = 0;
   for (i=0;i<points*2;i+=2) {
     // convert into device coordinates...
     graphicsToDeviceCoordinates(gfx, &vertices[i], &vertices[i+1]);
@@ -435,10 +435,10 @@ void graphicsFillPoly(JsGraphics *gfx, int points, short *vertices) {
     if (y>maxy) maxy=y;
   }
   if (miny<0) miny=0;
-  if (maxy>=gfx->data.height) maxy=(short)(gfx->data.height-1);
+  if (maxy>=gfx->data.height) maxy=(int)(gfx->data.height-1);
   short minx[gfx->data.height];
   short maxx[gfx->data.height];
-  short y;
+  int y;
   for (y=miny;y<=maxy;y++) {
     minx[y] = (short)(gfx->data.width);
     maxx[y] = -1;
@@ -454,7 +454,7 @@ void graphicsFillPoly(JsGraphics *gfx, int points, short *vertices) {
       if (minx[y]<0) minx[y]=0;
       if (maxx[y]>=gfx->data.width) maxx[y]=(short)(gfx->data.width-1);
       // try and expand the rect that we fill
-      short oldy = y;
+      int oldy = y;
       while (y<maxy && minx[y+1]==minx[oldy] && maxx[y+1]==maxx[oldy])
         y++;
       // actually fill
@@ -466,7 +466,7 @@ void graphicsFillPoly(JsGraphics *gfx, int points, short *vertices) {
 
 #ifndef NO_VECTOR_FONT
 // prints character, returns width
-unsigned int graphicsFillVectorChar(JsGraphics *gfx, short x1, short y1, short size, char ch) {
+unsigned int graphicsFillVectorChar(JsGraphics *gfx, int x1, int y1, int size, char ch) {
   // no need to modify coordinates as graphicsFillPoly does that
   if (size<0) return 0;
   if (ch<vectorFontOffset || ch-vectorFontOffset>=vectorFontCount) return 0;
@@ -497,7 +497,7 @@ unsigned int graphicsFillVectorChar(JsGraphics *gfx, short x1, short y1, short s
 }
 
 // returns the width of a character
-unsigned int graphicsVectorCharWidth(JsGraphics *gfx, unsigned short size, char ch) {
+unsigned int graphicsVectorCharWidth(JsGraphics *gfx, unsigned int size, char ch) {
   NOT_USED(gfx);
   if (size<0) return 0;
   if (ch<vectorFontOffset || ch-vectorFontOffset>=vectorFontCount) return 0;
@@ -507,7 +507,7 @@ unsigned int graphicsVectorCharWidth(JsGraphics *gfx, unsigned short size, char 
 #endif
 
 /// Draw a simple 1bpp image in foreground colour
-void graphicsDrawImage1bpp(JsGraphics *gfx, short x1, short y1, short width, short height, const unsigned char *pixelData) {
+void graphicsDrawImage1bpp(JsGraphics *gfx, int x1, int y1, int width, int height, const unsigned char *pixelData) {
   int pixel = 256|*(pixelData++);
   int x,y;
   for (y=y1;y<y1+height;y++) {
@@ -522,8 +522,8 @@ void graphicsDrawImage1bpp(JsGraphics *gfx, short x1, short y1, short width, sho
 /// Scroll the graphics device (in user coords). X>0 = to right, Y >0 = down
 void graphicsScroll(JsGraphics *gfx, int xdir, int ydir) {
   // Ensure we flip coordinate system if needed
-  short x1 = 0, y1 = 0;
-  short x2 = xdir, y2 = ydir;
+  int x1 = 0, y1 = 0;
+  int x2 = xdir, y2 = ydir;
   graphicsToDeviceCoordinates(gfx, &x1, &y1);
   graphicsToDeviceCoordinates(gfx, &x2, &y2);
   xdir = x2-x1;
