+++ libs/graphics/lcd_js.c
@@ -18,7 +18,7 @@
 #include "jsinteractive.h"
 
 
-void lcdSetPixel_JS(JsGraphics *gfx, short x, short y, unsigned int col) {
+void lcdSetPixel_JS(JsGraphics *gfx, int x, int y, unsigned int col) {
   // look up setPixel and execute it!
 //  JsVar *lcdProto = jsvObjectGetChild(gfx->graphicsVar, JSPARSE_PROTOTYPE_VAR, 0);
  // if (lcdProto) {
@@ -36,7 +36,7 @@ void lcdSetPixel_JS(JsGraphics *gfx, short x, short y, unsigned int col) {
 //  }
 }
 
-void  lcdFillRect_JS(struct JsGraphics *gfx, short x1, short y1, short x2, short y2) {
+void  lcdFillRect_JS(struct JsGraphics *gfx, int x1, int y1, int x2, int y2) {
   JsVar *fillRect = jsvObjectGetChild(gfx->graphicsVar/*lcdProto*/, "iFillRect", 0);
   if (fillRect) {
     JsVar *args[5];
